[LV] Layout View: Started refactoring of structure
Enclosed the canvas inside another div so that this div will have a height that is the rest of the screen's height minus the control's height. The canvas frame now has the height of the active device and not a relative one.

This achieves the effect of a full device view while still being vertically scrollable to see the whole device.