[Overlay] Fixed the leftover overlay problem
Before, when something moved in WordPress, like the menu on small devices, the overlays will stay where they are.

We solved this by making LayoutActions.frameChanged as an async event. Then we trigger the frameChanged event to adjust the position of the overlays. The reason why this didn't work before was frameChanged was sync. The position hasn't changed by the time it's triggered so nothing happened.