[D&D] We now let the DropRules to take care of business
We want components to behave uniquely when something is being dropped to them. So we let the DropRules take care of business.

Now we drop nodes inside an empty column.

We don't have a placeholder anymore. Instead we inject the drag subject right away, we just indicate that it's the drag subject.