[D&D] Droppable now checks if the node supports it.
The node should have statics that handle dragging. If it has a static method that fires when dragging on a specific area, then it will check if the mousecursor is on that area within the node and fires the method.

Layout Component does not listen to the dragging events anymore.