[D&D] Move and Reorder Columns
I encountered a perplexing bug that is caused by new object stores getting created when the node is moved. Since the nodes have new stores, the "changeable" behavior doesn't listen to changes anymore (the old store was destroyed).

The fix was to re-listen to the store when the component did update.