[UI] Rearranged css classes
Every component uses a class and not a specific ID.
The less files contain styling for components' classes, and not for the app.

The apps (e.g. themebuilder) are the ones responsible in putting the components together. So the CSS that targets IDs should be inside the app's main less file.