[UI] Topbar of the Layout view now stays on top
.ui-layout is now the scrollable so that the topbar stays on top of the Layout view.

We wrapped ui-layout inside a wrapper that has height 100%. The wrapper has 90px padding(or whatever will be the height of topbar and components) to push .ui-layout down. Then we put a negative margin on ui-layout to push it up so that what is shown is 100% of ui-layout. If we don't do this, ui-layout will show partial content because the content goes beyond the screen.