[FIX] The best way to resubscribe to Node changes is on newProps
Why? Because if we put the resubscribe at afterUpdate, it won't go in if only the parent was changed and there are no actual changes in the node. But by then, the original store has changed already. So we need to resubscribe everytime the component is re-instantiated. In this case the best place to resubscribe is at newProps.