[Refactor] Put detection of node info into a separate module
We took getNodeInfo and getCursorPosition from droppable and moved it into a separate module "nodeinfo.js". We want to use it for the resizable behavior.

We also refactored the way cursor position is detected. Instead of returning if the position is "true" for a certain area, we return an object that sets the position as true or false depending on the threshold. This is more efficient and allows more things to be done with the information.