[Resizable] Created Resizable behavior
If this behavior is applied to a component, the component will have "resize" controls. This means that the cursor will change to an ns-resize or ew-resize when the cursor is placed on the edges. The active control edges can be set using "Resizable.setResizeHandle".

WIP: Call the callback when handles are used.