[Styling] Added toggleStyle
Toggle Style merges the properties into the style if it doesn't exist or if it exists but the value is not equal. But if the value exists and is equal, then the property will be removed.

Also improved the naming convention. Styles and Properties are now differentiated. Styles are individual styles per selector, and it has CSS Properties. Stylesheet is a set of Styles for a particular media query. Stylesheets are all Stylesheets per media.