[Styling] Use a ternary operator instead of a method to return the class name
We have to enclose the ternary operation inside String so that it can be parsed properly by JSX. It returns false if not enclosed.

Introduced a "getCascade" method that is expected to return a Cascade object which will contain information on cascading properties from different media queries particularly on 'all'

The Cascade object is not created at this point.