[Styling] Color Picker is now implemented as a Component
It's a dropdown position relatively to the top, and fixed to the left.

In order to not use javascript to position the color picker, we used a CSS technique to make sure the picker is positioned just below the color input box, while at the same time, it uses a fixed positioning to put the picker to the left of the window.